Feature/mw 72 claim to verifier

+++ src/containers/MessageView/MessageView.tsx
@@ -1,35 +1,24 @@
-import * as sdk from '@kiltprotocol/prototype-sdk'
 import * as React from 'react'
 import { connect } from 'react-redux'
+
 import MessageDetailView from '../../components/MessageDetailView/MessageDetailView'
 import MessageListView from '../../components/MessageListView/MessageListView'
 import Modal, { ModalType } from '../../components/Modal/Modal'
-import attestationService from '../../services/AttestationService'
-import ContactRepository from '../../services/ContactRepository'
 import ErrorService from '../../services/ErrorService'
-import FeedbackService, { notifySuccess } from '../../services/FeedbackService'
+import FeedbackService from '../../services/FeedbackService'
 import MessageRepository from '../../services/MessageRepository'
-import * as Claims from '../../state/ducks/Claims'
 import * as Wallet from '../../state/ducks/Wallet'
-import * as Attestations from '../../state/ducks/Attestations'
-import { Contact } from '../../types/Contact'
+import { Message } from '../../types/Message'
 import {
-  Message,
-  MessageBodyType,
-  RequestAttestationForClaim,
-  ClaimMessageBody,
-  ApproveAttestationForClaim,
-} from '../../types/Message'
-import { BlockUi } from '../../types/UserFeedback'
+  BlockingNotification,
+  BlockUi,
+  NotificationType,
+} from '../../types/UserFeedback'
+
 import './MessageView.scss'
-import KiltAction from 'src/types/Action'
 
 interface Props {
   selectedIdentity?: Wallet.Entry
-  addAttestationToClaim: (
-    claimHash: string,
-    attestation: sdk.IAttestation
-  ) => void
 }
 
 interface State {
@@ -48,8 +37,6 @@ class MessageView extends React.Component<Props, State> {
     this.onDeleteMessage = this.onDeleteMessage.bind(this)
     this.onOpenMessage = this.onOpenMessage.bind(this)
     this.onCloseMessage = this.onCloseMessage.bind(this)
-    this.attestCurrentClaim = this.attestCurrentClaim.bind(this)
-    this.importAttestation = this.importAttestation.bind(this)
   }
 
   public render() {
@@ -69,6 +56,7 @@ class MessageView extends React.Component<Props, State> {
             ref={el => {
               this.messageModal = el
             }}
+            showOnInit={true}
             type={ModalType.BLANK}
             header={`Message from ${currentMessage.sender}`}
             onCancel={this.onCloseMessage}
@@ -77,9 +65,7 @@ class MessageView extends React.Component<Props, State> {
               message={currentMessage}
               onDelete={this.onDeleteMessage}
               onCancel={this.onCloseMessage}
-            >
-              {this.getMessageActions()}
-            </MessageDetailView>
+            />
           </Modal>
         )}
       </section>
@@ -98,40 +84,41 @@ class MessageView extends React.Component<Props, State> {
     }
   }
 
-  private onDeleteMessage(id: string) {
+  private onDeleteMessage(message: Message) {
     const { currentMessage } = this.state
-    MessageRepository.deleteByMessageId(id)
-      .then(() => {
-        this.fetchMessages()
-        if (currentMessage) {
-          this.onCloseMessage()
-        }
-      })
-      .catch(error => {
-        ErrorService.log({
-          error,
-          message: `Could not delete message ${id}`,
-          origin: 'MessageView.onDeleteMessage()',
-          type: 'ERROR.FETCH.DELETE',
-        })
-      })
-  }
 
-  private getMessageActions() {
-    const { currentMessage } = this.state
-    const messageBodyType: MessageBodyType | undefined =
-      currentMessage && currentMessage.body && currentMessage.body.type
-
-    switch (messageBodyType) {
-      case MessageBodyType.REQUEST_ATTESTATION_FOR_CLAIM:
-        return <button onClick={this.attestCurrentClaim}>Attest Claim</button>
-      case MessageBodyType.APPROVE_ATTESTATION_FOR_CLAIM:
-        return (
-          <button onClick={this.importAttestation}>Import Attestation</button>
-        )
-      default:
-        return ''
+    if (!message.id) {
+      return
     }
+
+    if (currentMessage) {
+      this.onCloseMessage()
+      this.fetchMessages()
+    }
+
+    FeedbackService.addBlockingNotification({
+      header: 'Are you sure?',
+      message: `Do you want to delete message '${message.id}' from '${
+        message.sender
+      }'?`,
+      modalType: ModalType.CONFIRM,
+      onConfirm: (notification: BlockingNotification) => {
+        MessageRepository.deleteByMessageId(message.id as string)
+          .then(() => {
+            this.fetchMessages()
+            notification.remove()
+          })
+          .catch(error => {
+            ErrorService.log({
+              error,
+              message: `Could not delete message ${message.id}`,
+              origin: 'MessageView.onDeleteMessage()',
+              type: 'ERROR.FETCH.DELETE',
+            })
+          })
+      },
+      type: NotificationType.INFO,
+    })
   }
 
   private onOpenMessage(message: Message) {
@@ -148,12 +135,7 @@ class MessageView extends React.Component<Props, State> {
   }
 
   private onCloseMessage() {
-    this.setState({
-      currentMessage: undefined,
-    })
-    if (this.messageModal) {
-      this.messageModal.hide()
-    }
+    this.setState({ currentMessage: undefined })
   }
 
   private fetchMessages() {
@@ -185,95 +167,6 @@ class MessageView extends React.Component<Props, State> {
       })
     }
   }
-
-  private async attestCurrentClaim() {
-    const { currentMessage } = this.state
-
-    if (!currentMessage) {
-      this.onCloseMessage()
-      return
-    }
-
-    const blockUi: BlockUi = FeedbackService.addBlockUi({
-      headline: 'Create attestation',
-    })
-
-    this.onCloseMessage()
-
-    ContactRepository.findByKey(currentMessage.senderKey)
-      .then((claimer: Contact) => {
-        const claimMessageBody: ClaimMessageBody = (currentMessage.body as RequestAttestationForClaim)
-          .content
-        const claim: sdk.IClaim = claimMessageBody.claim
-        attestationService
-          .attestClaim(claim)
-          .then(async attestation => {
-            attestationService.saveInStore({
-              attestation,
-              claimerAddress: claim.owner,
-              claimerAlias: claimer.name,
-              ctypeHash: claimMessageBody.claim.ctype,
-              ctypeName: claimMessageBody.cType.name,
-            } as Attestations.Entry)
-            await this.sendClaimAttestedMessage(attestation, claimer, claim)
-            if (currentMessage.id) {
-              this.onDeleteMessage(currentMessage.id)
-            }
-            this.fetchMessages()
-            blockUi.remove()
-            notifySuccess('Attestation created.\nMessage sent to claimer.')
-          })
-          .catch(error => {
-            blockUi.remove()
-            ErrorService.log({
-              error,
-              message: 'Unable to create and store attestation on blockchain',
-              origin: 'MessageView.attestCurrentClaim()',
-              type: 'ERROR.BLOCKCHAIN',
-            })
-          })
-      })
-      .catch(error => {
-        ErrorService.log({
-          error,
-          message: 'Could not retrieve claimer',
-          origin: 'MessageView.attestCurrentClaim()',
-          type: 'ERROR.FETCH.GET',
-        })
-      })
-  }
-
-  private async sendClaimAttestedMessage(
-    attestation: sdk.IAttestation,
-    claimer: Contact,
-    claim: sdk.IClaim
-  ): Promise<Message> {
-    const attestationMessageBody: ApproveAttestationForClaim = {
-      content: {
-        attestation,
-        claim,
-      },
-      type: MessageBodyType.APPROVE_ATTESTATION_FOR_CLAIM,
-    }
-    return MessageRepository.send(claimer, attestationMessageBody)
-  }
-
-  private importAttestation() {
-    const { addAttestationToClaim } = this.props
-    const { currentMessage } = this.state
-
-    if (currentMessage && currentMessage.body) {
-      const {
-        claim,
-        attestation,
-      } = (currentMessage.body as ApproveAttestationForClaim).content
-      addAttestationToClaim(claim.hash, attestation)
-      this.onCloseMessage()
-      if (currentMessage.id) {
-        this.onDeleteMessage(currentMessage.id)
-      }
-    }
-  }
 }
 
 const mapStateToProps = (state: { wallet: Wallet.ImmutableState }) => {
@@ -282,18 +175,4 @@ const mapStateToProps = (state: { wallet: Wallet.ImmutableState }) => {
   }
 }
 
-const mapDispatchToProps = (dispatch: (action: KiltAction) => void) => {
-  return {
-    addAttestationToClaim: (
-      claimHash: string,
-      attestation: sdk.IAttestation
-    ) => {
-      dispatch(Claims.Store.addAttestation(claimHash, attestation))
-    },
-  }
-}
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(MessageView)
+export default connect(mapStateToProps)(MessageView)
